refactor: seperate update logic to different service
